多模态情感冲突识别智能体/YA_MCPServer_MMInspector_Submission/pyproject.toml
[project]
name = "YA_MCPServer_MMInspector"
version = "0.1.0"
description = "Multimodal Conflict Inspector Agent based on MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "art>=6.5",
    "black>=25.9.0",
    "colorlog>=6.10.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.0",
    "pyyaml>=6.0.2",
    "ruff>=0.14.4",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "gradio>=4.0.0",
    "scikit-learn>=1.2.0",
    "requests>=2.32.5",
]
